refactor: Sidebarをwidgetsレイヤーに移行
- Sidebarをsrc/widgets/sidebar/uiに移動
- Storybookファイルもwidgetsレイヤーに移動
- 必要なコンポーネントをshared/ui publicAPIに追加
- 全ページでSidebarのインポートパスを更新
- FSD準拠性を確保し、レイヤー間の依存関係を整理
