+++ src/pages/posts/ui/PostDetailPageClient.tsx
@@ -3,7 +3,7 @@
 import { Post } from "@/entities/post";
 import type { PostType, User } from "@/shared/entities";
 import { sideBarWidthBase } from "@/shared/lib/const";
-import { Sidebar } from "@/shared/ui";
+import { Sidebar } from "@/widgets/sidebar";
 import { Box, Flex, Space } from "@mantine/core";
 import { useRouter } from "next/navigation";
