+++ src/pages/profile/ui/ProfilePageClient.tsx
@@ -2,8 +2,8 @@
 
 import type { PostType, ProfileType, User } from "@/shared/entities";
 import { sideBarWidthBase } from "@/shared/lib/const";
-import { Sidebar } from "@/shared/ui";
 import { PostsList } from "@/widgets/posts-list";
+import { Sidebar } from "@/widgets/sidebar";
 import { Box, Flex } from "@mantine/core";
 import { EditProfileModal } from "./EditProfileModal";
