+++ src/shared/ui/index.ts
@@ -1,7 +1,10 @@
-export { Sidebar } from "./Sidebar/Sidebar";
 export { ImageUpload } from "./ImageUpload/ImageUpload";
 export { NewPostForm } from "./NewPostForm/NewPostForm";
 export { CenteredLoader } from "./CenteredLoader/CenteredLoader";
 export { ContentPart, parseContent } from "./ContentPart/ContentPart";
 export { DateFormat } from "./DateFormat/DateFormat";
 export { PostActionsButtonGroup } from "./PostActionsButtonGroup/PostActionsButtonGroup";
+export { AccountButton } from "./AccountButton/AccountButton";
+export { LoginButton } from "./LoginButton/LoginButton";
+export { MainLinks } from "./MainLinks/MainLinks";
+export { PostModal } from "./Modal/PostModal/PostModal";
