+++ src/widgets/sidebar/ui/Sidebar.stories.tsx
